fix dev

+++ app/src/main/java/org/p2p/wallet/swap/ui/jupiter/main/JupiterSwapPresenter.kt
@@ -172,7 +172,7 @@ class JupiterSwapPresenter(
                     // todo also check for slippage error
                     Timber.e(result, "Failed to swap tokens")
                     val causeFailure = if (result.cause is JupiterSwapInteractor.LowSlippageRpcError) {
-                        TransactionStateSwapFailureReason.LowSlippage(currentState.slippage)
+                        TransactionStateSwapFailureReason.LowSlippage(currentState.slippage.doubleValue)
                     } else {
                         TransactionStateSwapFailureReason.Unknown(result.message.orEmpty())
                     }
